#4213
Vuestic's option "css" accepts an array of  css modules as an input. The default value of that option is ['smart-helpers', 'typography'].
The problem is that nuxt won't override the default array with an array passed by user (it's a known issue). Instaed two arrays will be concatenated.
For example, I want to use only a typography css module, so I set the css option to ['typography']. In that case useVuesticCSS method wil receive ['smart-helpers', 'typography', 'typography'] and not just ['typography']